+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+What's new in version 0.8
+-------------------------
+
+- Allowed ``show_percentages`` to be provided with a custom formatting string,
+  for example to show more decimal places. (:issue:`194`)
+
 What's new in version 0.7
 -------------------------